add concrete implmentations of length on all base collection types
